got searching working

+++ java/src/main/java/com/techelevator/dao/UserSqlDAO.java
@@ -214,7 +214,7 @@ public class UserSqlDAO implements UserDAO {
 
     private User mapRowToUser(SqlRowSet rs) {
         User user = new User();
-        user.setId(rs.getLong("user_id"));
+        user.setId(rs.getInt("user_id"));
         user.setUsername(rs.getString("username"));
         user.setPassword(rs.getString("password_hash"));
         user.setAuthorities(rs.getString("role"));
@@ -224,7 +224,7 @@ public class UserSqlDAO implements UserDAO {
     
     private User mapRowToUserPublic(SqlRowSet rs) {
         User user = new User();
-        user.setId(rs.getLong("user_id"));
+        user.setId(rs.getInt("user_id"));
         user.setUsername(rs.getString("username"));
         return user;
     }
